added respond func

+++ internal/model/error_response.go
@@ -1,7 +1,7 @@
 package model
 
 type ErrorResponse struct {
-	Code    int16  `json:"code"`
+	Code    int    `json:"code"`
 	Message string `json:"message"`
 	Error   error  `json:"error"`
 }
